+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Changes for the upcoming release can be found (and will be added on merging feat
 Changes from previous releases are listed below.
 
 ## Upcoming Release
-_No changes yet._
+
+- Add acquisition of Sentinel-1 data _(see #35)_
 
 ## 0.3.1 (2024-07-29)